# docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Sample API
  description: Demo REST API generated offline from Spring MVC annotations
  contact:
    name: API Team
    email: api@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
- url: https://api.example.com/sample-api/
  description: Production
- url: https://staging.example.com/sample-api/
  description: Staging
- url: http://localhost:8080/sample-api/
  description: Local development
security:
- bearerAuth: []
...
paths:
...
  /api/v1/agents:
    get:
      tags:
      - Generic REST API
      - Agents
      summary: Get all entities
      description: "Returns a paginated list of entities, with optional filtering\
        \ and sorting"
      operationId: getAll
      parameters:
      - name: page
        in: query
        description: Page number (0-based)
        required: true
        schema:
          type: integer
          format: int32
        example: "0"
      - name: size
        in: query
        description: Page size
        required: true
        schema:
          type: integer
          format: int32
        example: "20"
      responses:
        "200":
          description: Entities fetched successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentDto'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentDto'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentDto'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentDto'
    post:
      tags:
      - Generic REST API
      - Agents
      summary: Create new entity
      description: Creates a new instance of the entity
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentDto'
        required: true
      responses:
        "201":
          description: Entity created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDto'
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDto'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDto'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDto'
...
  /api/v1/agents/{id}/group:
    get:
      tags:
      - Generic REST API
      - Agents
      summary: Get Agent Group by Agent ID
      description: Returns the agent group associated with the given agent ID
      operationId: getAgentGroupByAgentId
      parameters:
      - name: id
        in: path
        description: Logical agent ID
        required: true
        schema:
          type: string
        example: agent-123
      responses:
        "200":
          description: Agent group found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentGroupDto'
        "404":
          description: Agent or group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentGroupDto'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentGroupDto'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentGroupDto'
  /api/v1/users:
    get:
      tags:
      - users
      summary: List users
      description: "Returns a paginated list of users, optionally filtered by name\
        \ or role."
      operationId: listUsers
      parameters:
      - name: name
        in: query
        description: Filter by name fragment (case-insensitive)
        required: false
        schema:
          type: string
        example: Maria
      - name: role
        in: query
        description: Filter by role
        required: false
        schema:
          type: string
        example: ADMIN
      - name: page
        in: query
        description: Page number (0-based)
        required: true
        schema:
          type: integer
          format: int32
        example: "0"
      - name: size
        in: query
        description: Page size (max 100)
        required: true
        schema:
          type: integer
          format: int32
        example: "20"
      responses:
        "200":
          description: Page of users returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponseUserDto'
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponseUserDto'
    post:
      tags:
      - users
      summary: Create user
      description: Registers a new user account. The email must be unique.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        "400":
          description: Validation error in request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
  /api/v1/users/{id}:
    get:
      tags:
      - users
      summary: Get user by ID
      description: Returns the full profile of a single user.
      operationId: getUserById
      parameters:
      - name: id
        in: path
        description: Unique identifier of the user
        required: true
        schema:
          type: integer
          format: int64
        example: "42"
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
    put:
      tags:
      - users
      summary: Update user
      description: Replaces all mutable fields of an existing user.
      operationId: updateUser
      parameters:
      - name: id
        in: path
        description: Unique identifier of the user to update
        required: true
        schema:
          type: integer
          format: int64
        example: "42"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
    delete:
      tags:
      - users
      summary: Delete user
      description: Permanently removes a user account.
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        description: Unique identifier of the user to delete
        required: true
        schema:
          type: integer
          format: int64
        example: "42"
      responses:
        "204":
          description: User deleted successfully
        "404":
          description: User not found
  /api/v1/users/search:
    get:
      tags:
      - users
      summary: Search users
      description: Full-text search across name and email fields.
      operationId: searchUsers
      parameters:
      - name: q
        in: query
        description: Search query string
        required: true
        schema:
          type: string
        example: silva
      - name: page
        in: query
        description: Page number (0-based)
        required: true
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        description: Page size
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponseUserDto'
...
components:
  schemas:
...
    AgentDto:
      required:
      - name
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the agent
          readOnly: true
          example: agent-123
        name:
          type: string
          description: Display name of the agent
          example: Alice Smith
        ssoId:
          type: string
          description: SSO identifier
          example: alice.smith@8x8.com
        ldapId:
          type: string
          description: LDAP identifier
          example: "cn=alice,ou=agents,dc=8x8,dc=com"
      description: Agent data transfer object
    AgentGroupDto:
      required:
      - name
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the agent group
          readOnly: true
          example: group-001
        name:
          type: string
          description: Display name of the agent group
          example: Tier-1 Support
      description: Agent group data transfer object
    PagedResponseUserDto:
      type: object
      properties:
        content:
          type: array
          description: Items in the current page
          items:
            $ref: '#/components/schemas/UserDto'
        page:
          type: integer
          description: Current page number (0-based)
          format: int32
          example: 0
        size:
          type: integer
          description: Number of items per page
          format: int32
          example: 20
        totalElements:
          type: integer
          description: Total number of items across all pages
          format: int64
          example: 150
        totalPages:
          type: integer
          description: Total number of pages
          format: int32
          example: 8
        last:
          type: boolean
          description: Whether this is the last page
          example: false
      description: Paginated response wrapper
    UserDto:
      required:
      - email
      - name
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier of the user
          format: int64
          readOnly: true
          example: 42
        name:
          type: string
          description: Full name of the user
          example: Maria Silva
        email:
          type: string
          description: Unique email address
          example: maria@example.com
        role:
          type: string
          description: Role assigned to the user
          example: ADMIN
          enum:
          - USER
          - ADMIN
          - MANAGER
        active:
          type: boolean
          description: Whether the account is active
          example: true
        createdAt:
          type: string
          description: Timestamp when the account was created
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          description: Timestamp of the last modification
          format: date-time
          readOnly: true
      description: User data transfer object
    CreateUserRequest:
      required:
      - email
      - name
      - password
      type: object
      properties:
        name:
          type: string
          description: Full name of the user
          example: João Costa
        email:
          type: string
          description: Unique email address
          example: joao@example.com
        password:
          type: string
          description: Plain-text password (will be hashed server-side)
          example: s3cr3t!
        role:
          type: string
          description: Role to assign to the user
          example: USER
          default: USER
          enum:
          - USER
          - ADMIN
          - MANAGER
      description: Payload for creating or updating a user
...
  securitySchemes:
    bearerAuth:
      type: http
      description: "JWT token obtained from the auth service (Authorization: Bearer\
        \ <token>)"
      scheme: bearer
      bearerFormat: JWT